app/requirements.txt
Django==4.2.5
django-cors-headers==4.2.0
django-split-settings==1.2.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
python-dotenv==1.0.0
jsonschema==4.19.0
uWSGI==2.0.22